pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clean-telegram"
version = "0.1.0"
description = "Script para limpar conta Telegram via Telethon"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Gabriel Ramos"}
]
keywords = ["telegram", "telethon", "cleanup", "automation", "privacy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]

dependencies = [
    "telethon>=1.42.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "qrcode>=8.0,<9.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
clean-telegram = "clean_telegram.__main__:main"

[project.urls]
Homepage = "https://github.com/gabrielramos/CleanTelegram"
Repository = "https://github.com/gabrielramos/CleanTelegram"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
clean_telegram = ["py.typed"]

# Black configuration
[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | \.eggs
    | \.tox
    | build
    | dist
    | __pycache__
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
src_paths = ["src", "tests"]
skip_gitignore = true

# MyPy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
show_error_codes = true
plugins = []

[[tool.mypy.overrides]]
module = "telethon.*"
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/clean_telegram",
    "--cov-report=term-missing",
    "--cov-report=html",
    "-v",
]
asyncio_mode = "auto"
markers = [
    "asyncio: mark test as async",
    "integration: mark test as integration test",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
branch = true
parallel = true

[tool.coverage.paths]
source = [
    "src/",
]

[tool.coverage.report]
fail_under = 55
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
]
